TST/REF: move data fixtures to separate file
